Added a check-mark on the input currently active for each display
When you open the menu, read the current input for each monitor and then
update the menu to add a check-mark for the active inputs. There will
be a small delay before this update to the menu items based on how
long it takes to run the ddcutil command(s).