Improved the "refresh" behaviour and fixed a bug
1. The refresh menu item will no longer close the menu. Instead it will
leave the menu open, remove the display items and add a
"detecting monitors..." item, then re-read "ddcutil detect". Once
the ddcutil command is done the newly detected monitors will be added
back to the method.

2. Leave the "options" and "refresh" menu items in place when errors
were detected or when monitor detection is in progress.

3. Fixed a bug where the default input check-mark would appear for all
the input menu items after a refresh in some cases.